+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Inspired by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Note: this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.2] - 2020-04-22
+### Fixed
+- Fixed bug where an `ArgumentError` would raise out of `Target#with_static_context` when a target hash uses a nested key that doesn't exist in the static context
+
 ## [0.8.1] - 2020-04-21
 ### Fixed
 - Fixed bug where `MonitorStub` was not responding to `#[]`